+++ src/editor/index.js
@@ -19,6 +19,7 @@ import { register, dispatch, select } from '@wordpress/data';
  */
 import { store as postKindsStore, STORE_NAME } from './stores/post-kinds';
 import KindSelectorPanel from './kind-selector';
+import './promote-panel';
 
 // Register the data store.
 register( postKindsStore );